refactor(edecimal): split edecimal into core / manipulators / iostream (#1334)
Phase 2 of the header layering, continuing after fixpnt (#1421).

    core.hpp          arithmetic, no streams          73,040 lines, 0 I/O-family
    manipulators.hpp  to_binary / to_string           77,152
    iostream.hpp      operator<< / operator>>         77,198
    edecimal.hpp      everything, as before           79,372

The umbrella was 98,017 lines pulling all five stream headers, so a translation unit
that only computes now reads 25% less and none of <iostream>, <sstream>, <iomanip>,
<ostream>, <istream>.

The type derives from std::vector<uint8_t>, so <vector> is structural in the core rather
than incidental; <string> stays for operator=(const std::string&) and parse(), the same
arithmetic surface integer's string constructor occupies. Two diagnostics that had been
holding <iostream> into the arithmetic path -- the "can this happen?" borrow check and
integer_divide_by_zero -- now use fprintf(stderr, ...), the Phase 0 idiom. The friend
declarations need only <iosfwd>. printStats() genuinely needs <ostream>, but it sits
behind the ALPHA EDECIMAL_OPERATIONS_COUNT switch, so the include went inside that guard
rather than into the core.

core.hpp takes native/ieee754_core.hpp rather than native/ieee754.hpp: convert_ieee754()
wants extractFields() and ieee754_parameter<>, which is the bit-manipulation half added
in 7b7ff3c0.

Two things fixed along the way that are not layering:

sqrt() reported a negative argument on EVERY call. The #else of the
EDECIMAL_THROW_ARITHMETIC_EXCEPTION guard was not inside the sign test, so
sqrt(edecimal(144)) returned 12 and still wrote "decimal_negative_sqrt_arg" to stderr.
The report is now inside `if (f.isneg())`, and the non-throwing path returns zero
instead of edecimal(std::sqrt(-x)), which was building from a NaN.

sqrt.hpp and mathlib.hpp were not self-contained -- they used edecimal with no include
that provides it, so they compiled only behind a prior include. Both now include
core.hpp. This is the same defect class as #1422; these two are in the subtree being
split, so they are fixed here rather than filed.

Also closed the IWYU gaps the hardened #1389 sweep found in this subtree: <cstdint>,
<cassert>, <vector> and <algorithm> in edecimal_impl.hpp, <string> in exceptions.hpp,
<cstdint>/<limits>/edecimal_fwd.hpp in numeric_limits.hpp. Sweep is now clean at 0 gaps
across all 10 headers.

edecimal_impl.hpp remains non-self-contained by design; it is a fragment assembled by
core.hpp, as fixpnt_impl.hpp, cfloat_impl.hpp and integer_impl.hpp already are.

Verified: both gates pass on gcc and clang; all 25 translation units in the tree that
mention edecimal build and run clean on both compilers; ASCII clean.

BEHAVIOUR-CHANGE: sqrt(edecimal) no longer writes "decimal_negative_sqrt_arg" to stderr
for non-negative arguments, and returns 0 rather than a NaN-derived value for negative
ones
